% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_fhd.R
\name{add_fhd}
\alias{add_fhd}
\title{plot_fhd}
\usage{
add_fhd(
  plot,
  fhd_data,
  id_col = "fhd_id",
  height_col = "height",
  draw_col = "draw_id",
  prob_col = "probability",
  plot_by_cov = NULL
)
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
