% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_select.R
\name{mod_data_select_server}
\alias{mod_data_select_server}
\title{data_select Server Functions}
\usage{
mod_data_select_server(
  id,
  nav_id = "main-nav",
  parent_session,
  metadata_tbl,
  restore_payload = NULL
)
}
\arguments{
\item{id}{Module ID}

\item{nav_id}{The ID of the parent navigation bar (e.g., "main-nav"). This is required to allow the module to control navigation between tabs.}

\item{parent_session}{The session object of the parent Shiny app. This is required to allow the module to control navigation between tabs.}

\item{metadata_tbl}{Data frame of all available FHD metadata.}

\item{restore_payload}{Optional reactive carrying a previously saved session payload.}
}
\description{
data_select Server Functions
}
